Move fixtures into conftest
